@@ -12,7 +12,6 @@ import android.util.Log;
 
 import java.io.File;
 import java.util.Arrays;
-import java.util.Comparator;
 
 public class BootReceiver extends BroadcastReceiver {
 
@@ -26,12 +25,7 @@ public class BootReceiver extends BroadcastReceiver {
         if (files == null) files = new File[0];
 
         // Sort files so that they get executed in a repeatable and logical order.
-        Arrays.sort(files, new Comparator<File>() {
-            @Override
-            public int compare(File f1, File f2) {
-                return f1.getName().compareTo(f2.getName());
-            }
-        });
+        Arrays.sort(files, (f1, f2) -> f1.getName().compareTo(f2.getName()));
 
         StringBuilder logMessage = new StringBuilder();
         for (File file : files) {
@@ -68,5 +62,4 @@ public class BootReceiver extends BroadcastReceiver {
         if (!file.canRead()) file.setReadable(true);
         if (!file.canExecute()) file.setExecutable(true);
     }
-
 }
